Move ParentColumn handlers out of workflow module
Since it does not contain anything special to workflows and can  be reused in
other models;
