+++ src/ggrc/converters/handlers.py
@@ -487,6 +487,47 @@ class CheckboxColumnHandler(ColumnHandler):
       current_app.logger.error(error)
 
 
+class ParentColumnHandler(ColumnHandler):
+
+  """ handler for directly mapped columns """
+
+  parent = None
+
+  def __init__(self, row_converter, key, **options):
+    super(ParentColumnHandler, self).__init__(row_converter, key, **options)
+
+  def parse_item(self):
+    """ get parent object """
+    # pylint: disable=protected-access
+    if self.raw_value == "":
+      self.add_error(errors.MISSING_VALUE_ERROR, column_name=self.display_name)
+      return None
+    slug = self.raw_value
+    obj = self.new_objects.get(self.parent, {}).get(slug)
+    if obj is None:
+      obj = self.parent.query.filter(self.parent.slug == slug).first()
+    if obj is None:
+      self.add_error(errors.UNKNOWN_OBJECT,
+                     object_type=self.parent._inflector.human_singular.title(),
+                     slug=slug)
+    return obj
+
+  def set_obj_attr(self):
+    super(ParentColumnHandler, self).set_obj_attr()
+    # inherit context
+    obj = self.row_converter.obj
+    parent = getattr(obj, self.key, None)
+    if parent is not None and \
+       hasattr(obj, "context_id") and \
+       hasattr(parent, "context_id") and \
+       parent.context_id is not None:
+      obj.context_id = parent.context_id
+
+  def get_value(self):
+    value = getattr(self.row_converter.obj, self.key, self.value)
+    return value.slug
+
+
 class ProgramColumnHandler(ColumnHandler):
 
   def parse_item(self):
